Simplifies org-roam-agenda by using org-roam-tag-add & -remove
Org-roam defines the functions org-roam-tag-add and
org-roam-tag-remove which modify the filetags of the node at point in
a colon-separated form. Using these functions instead of inserting
tags manually has the effects of

1. easy composition with interactive use of -tag-add and -tag-remove.
inserting a tag into a colon-separated list or creating one if needed
is tricky and has to consider several cases;

2. removing checks, since -tag-add and -tag-remove are
idempotent. Before, we had to check whether a "todo" tag exists before
inserting or deleting one, but the built-in functions are safe to use
in all cases.