Adds org-roam tag toggle function, removes spurious agenda error
Defines org-roam-tag-toggle, which takes ONE TAG string and either
adds or removes it from the org-roam node at point.

Stops an error message caused by org-roam-agenda-update-todo-tag. The
function org-roam-tag-remove is /almost/ idempotent, doing nothing if
a user tries to remove a nonexistent tag UNLESS there are no tags of
any kind remaining. If the "filetags" keyword of a node is empty it
throws a user error, so we check first.